Move base backend into folder
This re-organises the base backend a bit to more closely align with the
implementations.

Also avoids disruption due to changes in import paths due to the recent
Drawer refector
