+++ setup.cfg
@@ -121,6 +121,7 @@ warn_unused_ignores = True
 warn_unreachable = True
 no_implicit_optional = True
 disallow_untyped_defs = False
+disable_error_code = method-assign
 [mypy-_cffi_backend]
 ignore_missing_imports = True
 [mypy-cairocffi._generated.ffi]
